Fix build: Float pan in splitFisheye, modern Bluetooth option, drop dead shader fn
- LayoutEditorBar: compute splitFisheye pan as Float so it matches
  ViewOrientation's Float fields (was Double, caused a type error)
- OwletControl: .allowBluetooth -> .allowBluetoothHFP (renamed in the
  current SDK), clears the deprecation warning
- Shaders.metal: remove unused rotateRay() helper (no callers)

+++ Cribsight/Monitor/LayoutEditorBar.swift
@@ -135,7 +135,7 @@ struct LayoutEditorBar: View {
         var fishSlots: [LayoutSlot] = []
         for i in 0..<newCount {
             // Spread virtual-PTZ aims evenly from -0.9 (left) to +0.9 (right).
-            let panFrac = Double(i) / Double(max(newCount - 1, 1))
+            let panFrac = Float(i) / Float(max(newCount - 1, 1))
             let pan = -0.9 + panFrac * 1.8
             let orientation = ViewOrientation(pan: pan, tilt: 1.0, zoom: 1.8)
             fishSlots.append(LayoutSlot(
